Experiment with adding copts to BUILD file for 3.1. Resolve some clang errors

+++ ch3/3.1/3.1.cpp
@@ -46,7 +46,7 @@ int main()
 
 	cout << "Your final grade is " << setprecision(3)
 	     << 0.2 * midterm + 0.4 * final + 0.4 * sum / count
-	     << setprecision(prec) << endl;
+	     << setprecision(static_cast<int>(prec)) << endl;
 
 	return 0;
 }
